Create project and basic structure.
This commit creates the whole structure of the project and the first
program to generate keys. It's still not complete but a couple tests
suggest that it's working according to what is expected.

Also, a script to generate random prime numbers is included, along
with a sample of those prime numbers.